commiting before i start changing the user routs from :userNameto :id in case i break everything...

+++ routes/questions.js
@@ -15,14 +15,19 @@ router.get('/', async function(req, res, next) {
   }
 });
 
-
-//get question by id - since i have other get routes is this :id a problem?  id can only be a number so as long are other routes are all words is this safe?
+// at the moment i don't use :id or /count but i built them so i'll leave them here for later
 router.get('/:id', mustExist("id", "questions", "id"), async function(req, res, next) {
   console.log("getting a particular question")
   try {
     const {id} = req.params;
-    const question = await db(`SELECT * FROM questions WHERE id=${id};`)
-    res.send(question.data[0]);
+    //  this if clause checks whether the param you passed in this position is a number
+    //  if it is it processes it as an id, if it is not it allows the router to move on to try other routes (like /count)
+    if (isNaN(+id)){
+      next()
+    } else {
+      const question = await db(`SELECT * FROM questions WHERE id=${id};`)
+      res.send(question.data[0]);
+    }
   }catch (err) {
     res.status(500).send(err);
   }
@@ -42,6 +47,8 @@ router.get('/count', async function(req, res, next) {
 });
 
 
+// //get question by id, I also don't think I actually need this route but it's here. 
+// //I think there is a bug in it at the moment but i can't be bothered to fix it becasue I don't need the route yet...
 // router.get('/byID/:id', questionMustExist, async function(req, res, next) {
 //   console.log("getting a particular question")
 //   try {
@@ -51,11 +58,10 @@ router.get('/count', async function(req, res, next) {
 //   }catch (err) {
 //     res.status(500).send(err);
 //   }
-//   // res.send("nothing");
 // });
 
 
-//get 1 random question
+//get a specified number of random questions
 router.get('/random/:count', async function(req, res, next) {
   const {count} = req.params;
   try {
@@ -66,27 +72,36 @@ router.get('/random/:count', async function(req, res, next) {
   }
 });
 
+
+// post a question.  Must post the question in syntax that javascript can evaluate and calculate an answer for
+// otherwise the calculate answer guard function will reject it
 router.post('/', calculateAnswer, async function(req, res, next) {
   const {question} = req.body;
   const answer = req.params.answer;
   try {
+    // post:
     await db(`INSERT INTO questions (question, answer) values ("${question}", "${answer}");`)
+    // get the new question object back (i may not need this but i am toying with showing it on the screen after you post it):
     const resultObject = await db(`SELECT * FROM questions ORDER BY ID DESC LIMIT 1;`)
+    // add a message to the body of the result for future use - currently I show the user this message.
     resultObject.msg = `Question '${question}' with answer '${answer}' successfully added to database`
     res.send(resultObject);
-    // res.send({msg: `Question '${question}' with answer '${answer}' successfully added to database`});
   } catch (err){
     res.status(500).send(err)
   }
 });
 
+
+// delete a question.  so far I don't use this but in future i would like to make it possible the parent to curate the questions
 router.delete('/:id', mustExist("id", "questions", "id"), async function(req, res, next) {
   const {id} = req.params;
-  console.log(id)
+  if (isNaN(+id)){
+    next()
+  }
   try {
     await db(`DELETE FROM questions WHERE id=${id};`)
     res.send({msg: `Question with id '${id}' successfully deleted from database`});
-    // is there a way to send this without running a 'get' to get the question info? res.send({msg: `Question '${question}' with answer '${answer}' successfully deleted from database`});
+    // is there a way to send the following without running a 'get' to get the question info? res.send({msg: `Question '${question}' with answer '${answer}' successfully deleted from database`});
   } catch (err){
     console.log(err)
     res.status(500).send(err)
